backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Configuration
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# Utilities
email-validator==2.2.0
httpx==0.27.2

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# Code Quality
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Type Stubs
types-python-jose==3.3.4.20240106
types-passlib==1.7.7.20240819
